+++ src/ElapsedPeriod.php
@@ -32,28 +32,6 @@ final class ElapsedPeriod
     }
 
     /**
-     * @psalm-pure
-     *
-     * @throws \LogicException When using a period containing months or years
-     */
-    public static function ofPeriod(Period $period): self
-    {
-        if ($period->months() !== 0 || $period->years() !== 0) {
-            // a month or a year is not constant
-            throw new \LogicException('Months and years can not be converted to microseconds');
-        }
-
-        $milliseconds = Period\Value::day->milliseconds($period->days()) +
-            Period\Value::hour->milliseconds($period->hours()) +
-            Period\Value::minute->milliseconds($period->minutes()) +
-            Period\Value::second->milliseconds($period->seconds()) +
-            $period->milliseconds();
-        $milliseconds *= 1_000;
-
-        return new self($milliseconds + $period->microseconds());
-    }
-
-    /**
      * @return int<0, max>
      */
     public function milliseconds(): int